reworked lb ports update
