+++ code/iaas/model/src/main/java/io/cattle/platform/core/dao/NetworkDao.java
@@ -1,8 +1,10 @@
 package io.cattle.platform.core.dao;
 
+import io.cattle.platform.core.model.Instance;
 import io.cattle.platform.core.model.IpAddress;
 import io.cattle.platform.core.model.Network;
 import io.cattle.platform.core.model.Nic;
+import io.cattle.platform.core.model.Port;
 import io.cattle.platform.core.model.Subnet;
 
 import java.util.List;
@@ -27,5 +29,8 @@ public interface NetworkDao {
     List<Subnet> getSubnets(Network network);
 
     List<Long> findInstancesInUseByServiceDriver(Long id);
+
+    void updateInstancePorts(Instance instance, List<String> newPortDefs, List<Port> toCreate,
+            List<Port> toRemove, Map<String, Port> toRetain);
 
 }
