+++ code/iaas/service-discovery/server/src/main/java/io/cattle/platform/servicediscovery/process/LoadBalancerServiceUpdatePostListener.java
@@ -2,6 +2,8 @@ package io.cattle.platform.servicediscovery.process;
 
 import io.cattle.platform.core.constants.InstanceConstants;
 import io.cattle.platform.core.constants.ServiceConstants;
+import io.cattle.platform.core.dao.InstanceDao;
+import io.cattle.platform.core.dao.NetworkDao;
 import io.cattle.platform.core.model.Instance;
 import io.cattle.platform.core.model.Port;
 import io.cattle.platform.core.model.Service;
@@ -10,7 +12,6 @@ import io.cattle.platform.engine.handler.ProcessPostListener;
 import io.cattle.platform.engine.process.ProcessInstance;
 import io.cattle.platform.engine.process.ProcessState;
 import io.cattle.platform.json.JsonMapper;
-import io.cattle.platform.object.process.StandardProcess;
 import io.cattle.platform.process.common.handler.AbstractObjectProcessLogic;
 import io.cattle.platform.servicediscovery.api.dao.ServiceExposeMapDao;
 import io.cattle.platform.servicediscovery.api.util.ServiceDiscoveryUtil;
@@ -35,6 +36,12 @@ public class LoadBalancerServiceUpdatePostListener extends AbstractObjectProcess
     @Inject
     ServiceExposeMapDao expMapDao;
 
+    @Inject
+    NetworkDao ntwkDao;
+
+    @Inject
+    InstanceDao instanceDao;
+
     @Override
     public String[] getProcessNames() {
         return new String[] { ServiceConstants.PROCESS_SERVICE_UPDATE };
@@ -58,33 +65,50 @@ public class LoadBalancerServiceUpdatePostListener extends AbstractObjectProcess
 
         // only perform update when ports got changed
         Object oldObj = state.getData().get("old");
-        List<String> oldPortSpecs = new ArrayList<>();
+        List<String> oldPortDefs = new ArrayList<>();
         if (oldObj != null) {
             Map<String, Object> old = (Map<String, Object>) oldObj;
             if (old.containsKey(ServiceConstants.FIELD_LAUNCH_CONFIG)) {
                 Map<String, Object> oldLC = (Map<String, Object>) old.get(ServiceConstants.FIELD_LAUNCH_CONFIG);
                 if (oldLC.get(InstanceConstants.FIELD_PORTS) != null) {
-                    oldPortSpecs = (List<String>) oldLC.get(InstanceConstants.FIELD_PORTS);
+                    oldPortDefs = (List<String>) oldLC.get(InstanceConstants.FIELD_PORTS);
                 }
             }
         }
 
         Map<String, Object> launchConfigData = ServiceDiscoveryUtil.getLaunchConfigDataAsMap(service,
                 ServiceConstants.PRIMARY_LAUNCH_CONFIG_NAME);
-        List<String> portSpecs = new ArrayList<>();
+        List<String> newPortDefs = new ArrayList<>();
         if (launchConfigData.get(InstanceConstants.FIELD_PORTS) != null) {
-            portSpecs = (List<String>) launchConfigData.get(InstanceConstants.FIELD_PORTS);
+            newPortDefs = (List<String>) launchConfigData.get(InstanceConstants.FIELD_PORTS);
         }
 
-        if (portSpecs.containsAll(oldPortSpecs) && oldPortSpecs.containsAll(portSpecs)) {
+        if (newPortDefs.containsAll(oldPortDefs) && oldPortDefs.containsAll(newPortDefs)) {
             return;
         }
 
         List<? extends Instance> serviceContainers = expMapDao.listServiceManagedInstancesAll(service);
         for (Instance instance : serviceContainers) {
-            instance = objectManager.setFields(instance, InstanceConstants.FIELD_PORTS, portSpecs);
-            objectProcessManager.scheduleStandardProcess(StandardProcess.UPDATE, instance,
-                    new HashMap<String, Object>());
+            List<Port> toCreate = new ArrayList<>();
+            List<Port> toRemove = new ArrayList<>();
+            Map<String, Port> toRetain = new HashMap<>();
+            // orchestrate port creation/removal
+            ntwkDao.updateInstancePorts(instance, newPortDefs, toCreate, toRemove, toRetain);
+            for (Port port : toCreate) {
+                port = objectManager.create(port);
+            }
+
+            for (Port port : toRetain.values()) {
+                createThenActivate(port, new HashMap<String, Object>());
+            }
+
+            for (Port port : toRemove) {
+                deactivateThenRemove(port, new HashMap<String, Object>());
+            }
+
+            // trigger instance/metadata update
+            instance = objectManager.setFields(instance, InstanceConstants.FIELD_PORTS, newPortDefs);
+            instanceDao.clearCacheInstanceData(instance.getId());
         }
     }
